activity7/requirements.txt
functions-framework==3.*
PyMySQL
python-dotenv